Created generic serial interface
A generic serial interface is created for early kernel development in
which a serial console is useful. The specific serial peripheral (the
PL011) has been abstracted out to its own implementation. The system
console can be set with a cargo feature, and the serial interfaces to
compile can also be set with a cargo feature